verdict-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
anthropic==0.40.0
elevenlabs==1.13.5
httpx==0.28.1
python-dotenv==1.0.1
nanoid==2.0.0
boto3==1.35.0
